+++ embedding-stores/langchain4j-community-cloud-sql-pg/src/main/java/dev/langchain4j/community/store/embedding/cloudsql/index/query/QueryOptions.java
@@ -1,0 +1,16 @@
+package dev.langchain4j.community.store.embedding.cloudsql.index.query;
+
+import java.util.List;
+
+/**
+ * Query options interface
+ */
+public interface QueryOptions {
+
+    /**
+     * Convert index attributes to list of configuration
+     *
+     * @return List of parameter setting strings
+     */
+    List<String> getParameterSettings();
+}
