--- embedding-stores/langchain4j-community-cloud-sql-pg/src/main/java/dev/langchain4j/store/embedding/index/query/QueryOptions.java
@@ -1,16 +1,0 @@
-package dev.langchain4j.store.embedding.index.query;
-
-import java.util.List;
-
-/**
- * Query options interface
- */
-public interface QueryOptions {
-
-    /**
-     * Convert index attributes to list of configuration
-     *
-     * @return List of parameter setting strings
-     */
-    List<String> getParameterSettings();
-}
